Fix issue #5 and #14
Fix architv/soccer-cli/5
Fix architv/soccer-cli/14

This replaces the .format() with the older % formatting.
The older method makes it easier to work with str and unicode at the
same time.

Since the % formatting is used elsewhere in the script I thought it
shouldn't really matter unless the newer .format() is really desired.