Fix module bin install issue
Fix architv/soccer-cli#16

Previously the setup script places 'main.py', 'leagueids.py' etc into
the /bin folder along with the 'soccer' script.

This was not desired as it meant not only was the `soccer` command
available via the command line (nice!)  but typing `main.py` or `leagueids.py`
would also run the python scripts...

This change means that only `soccer` gets placed in the /bin folder.