Fix csv unicode issue
Previously it would not work when one of the teams contained a special
unicode character.

This simply echos a unicode string.

And for writing via the csv module, it will go through each element
converting it to a unicode string with appropriate encoding.

This means we get the proper output on Windows too.