Fix parameter bug: soccer --league CL --standings
CL standings are not supported. They could be. But the API data is
different from other league tables. (Contains 8 separate league tables).
You would need to make a special case when dealing with CL. And since it
is a cup, the standings (group stage tables) are not that significant.

Also removed e.message since that is deprecated. It is removed
completely in Python 3.